pyproject.toml
[project]
name = "gev-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "gevla"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai (>=1.28.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
